requirements/prod.txt
python-freeipa>=1.0
python-dotenv>=0.20
